# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eyva OpenApi Documentation",
        "description": "Eyva Swagger OpenApi Documentation",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api/v1",
            "description": "Eyva API Server"
        }
    ],
    "paths": {
        "/abonnements": {
            "get": {
                "tags": [
                    "abonnements"
                ],
                "summary": "Get abonnements",
                "description": "Get abonnements",
                "operationId": "abonnements",
                "responses": {
                    "200": {
                        "description": "HTTP_OK"
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST"
                    },
                    "401": {
                        "description": "HTTP_UNAUTHORIZED"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/abonnements/{id}": {
            "get": {
                "tags": [
                    "abonnements"
                ],
                "summary": "Get one by id from abonnements",
                "description": "Get one by id from abonnements",
                "operationId": "showabonnements",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTTP_OK"
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST"
                    },
                    "401": {
                        "description": "HTTP_UNAUTHORIZED"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/abonnements/{abonnement_id}/buy": {
            "post": {
                "tags": [
                    "abonnements"
                ],
                "summary": "buy",
                "description": "buy",
                "operationId": "buyabonnements",
                "parameters": [
                    {
                        "name": "abonnement_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTTP_OK"
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST"
                    },
                    "401": {
                        "description": "HTTP_UNAUTHORIZED"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "login",
                "description": "login",
                "operationId": "login",
                "requestBody": {
                    "description": "Data for login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "default": "77012345678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "default": "Password1"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "default": "Password1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "HTTP_OK"
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST"
                    },
                    "401": {
                        "description": "HTTP_UNAUTHORIZED"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "refresh",
                "description": "refresh",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "HTTP_OK"
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST"
                    },
                    "401": {
                        "description": "HTTP_UNAUTHORIZED"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "logout",
                "description": "logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "HTTP_OK"
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST"
                    },
                    "401": {
                        "description": "HTTP_UNAUTHORIZED"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/profile/transactions": {
            "get": {
                "tags": [
                    "profile"
                ],
                "summary": "Get transactions",
                "description": "Get transactions",
                "operationId": "transactions",
                "responses": {
                    "200": {
                        "description": "HTTP_OK"
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST"
                    },
                    "401": {
                        "description": "HTTP_UNAUTHORIZED"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/services": {
            "get": {
                "tags": [
                    "services"
                ],
                "summary": "Get services",
                "description": "Get services",
                "operationId": "services",
                "responses": {
                    "200": {
                        "description": "HTTP_OK"
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST"
                    },
                    "401": {
                        "description": "HTTP_UNAUTHORIZED"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with phone and password to get the authentication token",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}